@@ -9,30 +9,52 @@ namespace WeCantSpell.Hunspell;
 
 public readonly struct BreakSet : IReadOnlyList<string>
 {
-    public static BreakSet Empty { get; } = new(Array.Empty<string>());
+    public static BreakSet Empty { get; } = new([]);
 
-    public static BreakSet Create(IEnumerable<string> entries) =>
-        new((entries ?? throw new ArgumentNullException(nameof(entries))).ToArray());
+    public static BreakSet Create(IEnumerable<string> entries)
+    {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(entries);
+#else
+        if (entries is null) throw new ArgumentNullException(nameof(entries));
+#endif
+
+        return new(entries.ToArray());
+    }
 
     internal BreakSet(string[] entries)
     {
-#if DEBUG
-        if (entries is null) throw new ArgumentNullException(nameof(entries));
-#endif
         _entries = entries;
     }
 
-    private readonly string[] _entries;
+    private readonly string[]? _entries;
 
-    public int Count => _entries.Length;
+    public int Count => (_entries?.Length).GetValueOrDefault();
+
     public bool IsEmpty => !HasItems;
+
     public bool HasItems => _entries is { Length: > 0 };
-    public string this[int index] => _entries[index];
 
-    public IEnumerator<string> GetEnumerator() => ((IEnumerable<string>)_entries).GetEnumerator();
-    IEnumerator IEnumerable.GetEnumerator() => _entries.GetEnumerator();
+    public string this[int index]
+    {
+        get
+        {
+#if HAS_THROWOOR
+            ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+            if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
 
-    internal string[] GetInternalArray() => _entries;
+            return _entries![index];
+        }
+    }
+
+    public IEnumerator<string> GetEnumerator() => ((IEnumerable<string>)GetInternalArray()).GetEnumerator();
+
+    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+    internal string[] GetInternalArray() => _entries ?? [];
 
     /// <summary>
     /// Calculate break points for recursion limit.
@@ -41,9 +63,9 @@ public readonly struct BreakSet : IReadOnlyList<string>
     {
         var nbr = 0;
 
-        if (scw.Length != 0)
+        if (scw.Length != 0 && HasItems)
         {
-            foreach (var breakEntry in _entries)
+            foreach (var breakEntry in _entries!)
             {
                 var pos = 0;
                 while ((pos = scw.IndexOf(breakEntry, pos, StringComparison.Ordinal)) >= 0)
@@ -64,9 +86,9 @@ public readonly struct BreakSet : IReadOnlyList<string>
     {
         var nbr = 0;
 
-        if (scw is { Length: > 0 })
+        if (scw.Length != 0 && HasItems)
         {
-            foreach (var breakEntry in _entries)
+            foreach (var breakEntry in _entries!)
             {
                 var pos = 0;
                 while ((pos = scw.IndexOf(breakEntry, pos, StringComparison.Ordinal)) >= 0)
