@@ -1,14 +1,15 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Threading;
 
 namespace WeCantSpell.Hunspell.Infrastructure;
 
-public sealed class ArrayBuilder<T>
+sealed class ArrayBuilder<T> : IList<T>
 {
     public ArrayBuilder()
     {
-        _values = Array.Empty<T>();
+        _values = [];
     }
 
     internal ArrayBuilder(T[] values)
@@ -19,10 +20,13 @@ public sealed class ArrayBuilder<T>
 
     public ArrayBuilder(int initialCapacity)
     {
-#if DEBUG
+#if HAS_THROWOOR
+        ArgumentOutOfRangeException.ThrowIfLessThan(initialCapacity, 0);
+#else
         if (initialCapacity < 0) throw new ArgumentOutOfRangeException(nameof(initialCapacity));
 #endif
-        _values = initialCapacity == 0 ? Array.Empty<T>() : new T[initialCapacity];
+
+        _values = initialCapacity == 0 ? [] : new T[initialCapacity];
     }
 
     private T[] _values;
@@ -31,6 +35,8 @@ public sealed class ArrayBuilder<T>
 
     public int Capacity => _values.Length;
 
+    public bool IsReadOnly => false;
+
     public void Clear()
     {
         Count = 0;
@@ -38,17 +44,43 @@ public sealed class ArrayBuilder<T>
 
     public T this[int index]
     {
-        get => index >= 0 && index < Count ? _values[index] : throw new ArgumentOutOfRangeException(nameof(index));
+        get
+        {
+#if HAS_THROWOOR
+            ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+            if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
+            return _values[index];
+        }
         set
         {
-            if (index < 0 || index >= Count)
-            {
-                throw new ArgumentOutOfRangeException(nameof(index));
-            }
-
+#if HAS_THROWOOR
+            ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+            if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
             _values[index] = value;
         }
     }
+
+    public int IndexOf(T item) => Array.IndexOf(_values, item, 0, Count);
+
+    public bool Contains(T item) => IndexOf(item) >= 0;
+
+    public void CopyTo(T[] array, int arrayIndex) => Array.Copy(_values, 0, array, arrayIndex, Count);
+
+    public IEnumerator<T> GetEnumerator()
+    {
+        for (var i = 0; i < Count; i++)
+        {
+            yield return _values[i];
+        }
+    }
+
+    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
     public void Add(T value)
     {
@@ -62,7 +94,11 @@ public sealed class ArrayBuilder<T>
 
     public void AddRange(IEnumerable<T> values)
     {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(values);
+#else
         if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
 
         foreach (var value in values)
         {
@@ -72,7 +108,11 @@ public sealed class ArrayBuilder<T>
 
     public void AddRange(ICollection<T> values)
     {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(values);
+#else
         if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
 
         var futureSize = Count + values.Count;
         EnsureCapacityAtLeast(futureSize);
@@ -90,7 +130,11 @@ public sealed class ArrayBuilder<T>
 
     public void AddRange(T[] values)
     {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(values);
+#else
         if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
 
         var futureSize = Count + values.Length;
         EnsureCapacityAtLeast(futureSize);
@@ -98,23 +142,46 @@ public sealed class ArrayBuilder<T>
         Count = futureSize;
     }
 
+    public void AddRange(ReadOnlySpan<T> values)
+    {
+        var futureSize = Count + values.Length;
+        EnsureCapacityAtLeast(futureSize);
+        values.CopyTo(_values.AsSpan());
+        Count = futureSize;
+    }
+
     public void GrowToCapacity(int requiredLength)
     {
+#if HAS_THROWOOR
+        ArgumentOutOfRangeException.ThrowIfLessThan(requiredLength, 0);
+#else
         if (requiredLength < 0) throw new ArgumentOutOfRangeException(nameof(requiredLength));
+#endif
 
         if (_values.Length < requiredLength)
         {
-            Array.Resize(ref _values, requiredLength);
+            if (Count == 0)
+            {
+                _values = new T[requiredLength];
+            }
+            else
+            {
+                Array.Resize(ref _values, requiredLength);
+            }
         }
     }
 
     public void EnsureCapacityAtLeast(int requiredLength)
     {
+#if HAS_THROWOOR
+        ArgumentOutOfRangeException.ThrowIfLessThan(requiredLength, 0);
+#else
         if (requiredLength < 0) throw new ArgumentOutOfRangeException(nameof(requiredLength));
+#endif
 
         if (_values.Length < requiredLength)
         {
-            Array.Resize(ref _values, CalculateBestCapacity(requiredLength));
+            GrowToCapacity(CalculateBestCapacity(requiredLength));
         }
     }
 
@@ -136,16 +203,6 @@ public sealed class ArrayBuilder<T>
 
         Insert(insertLocation, value);
         return true;
-    }
-
-    internal int BinarySearch(int startIndex, int count, T value)
-    {
-#if DEBUG
-        if (startIndex < 0 || startIndex >= Count) throw new ArgumentOutOfRangeException(nameof(startIndex));
-        if (startIndex + count > Count) throw new ArgumentOutOfRangeException(nameof(count));
-#endif
-
-        return Array.BinarySearch(_values, startIndex, count, value);
     }
 
     public void Insert(int insertionIndex, T value)
@@ -182,6 +239,42 @@ public sealed class ArrayBuilder<T>
         Count++;
     }
 
+    public void RemoveAt(int index)
+    {
+#if HAS_THROWOOR
+        ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+        ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+        if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
+
+        if (Count > 0)
+        {
+            var newSize = Count - 1;
+
+            if (index < newSize)
+            {
+                Array.Copy(_values, index + 1, _values, index, newSize - index);
+            }
+
+            _values[newSize] = default!;
+
+            Count = newSize;
+        }
+    }
+
+    public bool Remove(T item)
+    {
+        var index = IndexOf(item);
+        if (index >= 0)
+        {
+            RemoveAt(index);
+            return true;
+        }
+
+        return false;
+    }
+
     public T[] Extract()
     {
 #if DEBUG
@@ -205,9 +298,11 @@ public sealed class ArrayBuilder<T>
         }
     }
 
-    public T[] MakeArray() => Count == 0 ? Array.Empty<T>() : _values.AsSpan(0, Count).ToArray();
+    public T[] MakeArray() => Count == 0 ? [] : _values.AsSpan(0, Count).ToArray();
 
     public T[] MakeOrExtractArray(bool extract) => extract ? Extract() : MakeArray();
+
+    internal int BinarySearch(int startIndex, int count, T value) => Array.BinarySearch(_values, startIndex, count, value);
 
     private int CalculateBestCapacity(int minCapacity) => Math.Max(CalculateNextCapacity(), minCapacity);
 
@@ -215,7 +310,7 @@ public sealed class ArrayBuilder<T>
 
     internal static class Pool
     {
-        private const int MaxCapacity = 20;
+        private const int MaxCapacity = 32; // NOTE: Because of the growth values, this should be a power of two
 
         private static ArrayBuilder<T>? Cache;
 
@@ -227,7 +322,7 @@ public sealed class ArrayBuilder<T>
             }
             else
             {
-                taken = new();
+                taken = [];
             }
 
             return taken;
@@ -235,10 +330,6 @@ public sealed class ArrayBuilder<T>
 
         public static ArrayBuilder<T> Get(int capacity)
         {
-#if DEBUG
-            if (capacity <= 0) throw new ArgumentOutOfRangeException(nameof(capacity));
-#endif
-
             if (Interlocked.Exchange(ref Cache, null) is { } taken)
             {
                 taken.Clear();
@@ -254,19 +345,15 @@ public sealed class ArrayBuilder<T>
 
         public static void Return(ArrayBuilder<T> builder)
         {
-#if DEBUG
-            if (builder is null) throw new ArgumentNullException(nameof(builder));
-#endif
-
-            if (builder.Capacity > 0 && builder.Capacity <= MaxCapacity)
+            if (builder is { Capacity: <= MaxCapacity })
             {
                 Volatile.Write(ref Cache, builder);
             }
         }
 
-        public static T[] GetArrayAndReturn(ArrayBuilder<T> builder)
+        public static T[] ExtractAndReturn(ArrayBuilder<T> builder)
         {
-            var result = builder.MakeArray();
+            var result = builder.Extract();
             Return(builder);
             return result;
         }
