@@ -9,42 +9,61 @@ namespace WeCantSpell.Hunspell;
 
 public readonly struct PatternSet : IReadOnlyList<PatternEntry>
 {
-    public static PatternSet Empty { get; } = new(Array.Empty<PatternEntry>());
+    public static PatternSet Empty { get; } = new([]);
 
-    public static PatternSet Create(IEnumerable<PatternEntry> entries) =>
-        new((entries ?? throw new ArgumentNullException(nameof(entries))).ToArray());
+    public static PatternSet Create(IEnumerable<PatternEntry> entries)
+    {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(entries);
+#else
+        if (entries is null) throw new ArgumentNullException(nameof(entries));
+#endif
+
+        return new(entries.ToArray());
+    }
 
     internal PatternSet(PatternEntry[] patterns)
     {
-#if DEBUG
-        if (patterns is null) throw new ArgumentNullException(nameof(patterns));
-#endif
-
         _patterns = patterns;
     }
 
-    private readonly PatternEntry[] _patterns;
+    private readonly PatternEntry[]? _patterns;
 
-    public int Count => _patterns.Length;
+    public int Count => (_patterns?.Length).GetValueOrDefault();
+
     public bool IsEmpty => !HasItems;
+
     public bool HasItems => _patterns is { Length: > 0 };
-    public PatternEntry this[int index] => _patterns[index];
-    public IEnumerator<PatternEntry> GetEnumerator() => ((IEnumerable<PatternEntry>)_patterns).GetEnumerator();
-    IEnumerator IEnumerable.GetEnumerator() => _patterns.GetEnumerator();
+
+    public PatternEntry this[int index]
+    {
+        get
+        {
+#if HAS_THROWOOR
+            ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+            if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
+
+            return _patterns![index];
+        }
+    }
+
+    public IEnumerator<PatternEntry> GetEnumerator() => ((IEnumerable<PatternEntry>)GetInternalArray()).GetEnumerator();
+
+    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+    internal PatternEntry[] GetInternalArray() => _patterns ?? [];
 
     /// <summary>
     /// Forbid compoundings when there are special patterns at word bound.
     /// </summary>
     internal bool Check(ReadOnlySpan<char> word, int pos, WordEntry r1, WordEntry r2, bool affixed)
     {
-#if DEBUG
-        if (r1 is null) throw new ArgumentNullException(nameof(r1));
-        if (r2 is null) throw new ArgumentNullException(nameof(r2));
-#endif
-
         var wordAfterPos = word.Slice(pos);
 
-        foreach (var patternEntry in _patterns)
+        foreach (var patternEntry in GetInternalArray())
         {
             if (
                 (
