@@ -1,14 +1,15 @@
 ﻿using System.Collections.Generic;
 
 using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
+
 using WeCantSpell.Hunspell.Benchmarks.Helpers;
 
 namespace WeCantSpell.Hunspell.Benchmarks.Suites;
 
-[SimpleJob(id: "Check en-US")]
-[MinColumn, MaxColumn, MeanColumn, MedianColumn]
-[MinWarmupCount(1), MaxWarmupCount(5)]
-[MinIterationCount(1), MaxIterationCount(20), MinInvokeCount(1), IterationTime(250)]
+[SimpleJob(id: "Check en-US", runtimeMoniker: RuntimeMoniker.Net80, baseline: true)]
+[SimpleJob(id: "Check en-US", runtimeMoniker: RuntimeMoniker.Net60)]
+[MinColumn, MeanColumn, MedianColumn]
 public class CheckEnUsSuite
 {
     protected WordList WordList;
