@@ -1,11 +1,13 @@
 ﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Jobs;
+
 using WeCantSpell.Hunspell.Benchmarks.Helpers;
 
 namespace WeCantSpell.Hunspell.Benchmarks.Suites;
 
-[SimpleJob(id: "Suggest en-US")]
-[MinWarmupCount(1), MaxWarmupCount(5)]
-[MinIterationCount(1), MaxIterationCount(20), MinInvokeCount(1), IterationTime(250)]
+[SimpleJob(id: "Suggest en-US", runtimeMoniker: RuntimeMoniker.Net80, baseline: true)]
+[SimpleJob(id: "Suggest en-US", runtimeMoniker: RuntimeMoniker.Net60)]
+[MinColumn, MeanColumn, MedianColumn]
 public class SuggestEnUsSuite
 {
     protected WordList WordList;
