@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Buffers;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,17 +10,26 @@ namespace WeCantSpell.Hunspell;
 
 public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterSet>
 {
-    public static readonly CharacterSet Empty = new(Array.Empty<char>(), default);
+    public static readonly CharacterSet Empty =
+#if HAS_SEARCHVALUES
+        new([]);
+#else
+        new([], default);
+#endif
 
     public static bool operator ==(CharacterSet left, CharacterSet right) => left.Equals(right);
 
     public static bool operator !=(CharacterSet left, CharacterSet right) => !(left == right);
 
-    public static CharacterSet Create(char value) => new(new[] { value }, value);
+    public static CharacterSet Create(char value) => new(value);
 
     public static CharacterSet Create(IEnumerable<char> values)
     {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(values);
+#else
         if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
 
         var builder = new Builder();
         builder.AddRange(values);
@@ -28,7 +38,11 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
 
     public static CharacterSet Create(string values)
     {
+#if HAS_THROWNULL
+        ArgumentNullException.ThrowIfNull(values);
+#else
         if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
 
         return Create(values.AsSpan());
     }
@@ -45,21 +59,68 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
         return builder.Create(allowDestructive: true);
     }
 
-    private CharacterSet(char[] values, char mask)
+#if HAS_SEARCHVALUES
+
+    private CharacterSet(char value) : this([value])
     {
-        _mask = mask;
-        _values = values;
     }
 
-    private readonly char[] _values;
+    private CharacterSet(char[] values)
+    {
+        _values = values;
+        _searchValues = SearchValues.Create(values);
+    }
+
+    private readonly char[]? _values;
+    private readonly SearchValues<char>? _searchValues;
+
+#else
+
+    private CharacterSet(char value) : this([value], value)
+    {
+    }
+
+    private CharacterSet(char[] values, char mask)
+    {
+        _values = values;
+        _mask = mask;
+    }
+
+    private readonly char[]? _values;
     private readonly char _mask;
 
-    public int Count => _values.Length;
+#endif
+
+    public int Count => (_values?.Length).GetValueOrDefault();
     public bool IsEmpty => !HasItems;
     public bool HasItems => _values is { Length: > 0 };
-    public char this[int index] => _values[index];
-    public IEnumerator<char> GetEnumerator() => ((IEnumerable<char>)_values).GetEnumerator();
-    IEnumerator IEnumerable.GetEnumerator() => _values.GetEnumerator();
+    public char this[int index]
+    {
+        get
+        {
+#if HAS_THROWOOR
+            ArgumentOutOfRangeException.ThrowIfLessThan(index, 0);
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(index, Count);
+#else
+            if (index < 0 || index >= Count) throw new ArgumentOutOfRangeException(nameof(index));
+#endif
+            return _values![index];
+        }
+    }
+
+    public IEnumerator<char> GetEnumerator() => ((IEnumerable<char>)GetInternalArray()).GetEnumerator();
+
+    IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+    private char[] GetInternalArray() => _values ?? [];
+
+#if HAS_SEARCHVALUES
+
+    public bool Contains(char value) => (_searchValues?.Contains(value)).GetValueOrDefault();
+
+    public int FindIndexOfMatch(ReadOnlySpan<char> text) => _searchValues is not null ? text.IndexOfAny(_searchValues) : -1;
+
+#else
 
     public bool Contains(char value)
     {
@@ -123,6 +184,8 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
 
         return -1;
     }
+
+#endif
 
     public int FindIndexOfMatch(ReadOnlySpan<char> text, int startIndex)
     {
@@ -195,13 +258,18 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
         return StringBuilderPool.GetStringAndReturn(builder).AsSpan();
     }
 
-    public override string ToString() => new(_values);
+    public override string ToString() => new(GetInternalArray());
 
-    public bool Equals(CharacterSet obj) => _values.SequenceEqual(obj._values);
+    public bool Equals(CharacterSet obj) => GetInternalArray().SequenceEqual(obj.GetInternalArray());
 
     public override bool Equals(object? obj) => obj is CharacterSet set && Equals(set);
 
-    public override int GetHashCode() => HashCode.Combine(Count, _mask);
+    public override int GetHashCode() =>
+#if HAS_SEARCHVALUES
+        HashCode.Combine(Count, HasItems ? _values![0] : default);
+#else
+        HashCode.Combine(Count, _mask);
+#endif
 
     public sealed class Builder
     {
@@ -218,25 +286,32 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
         }
 
         private readonly ArrayBuilder<char> _builder;
+
+#if HAS_SEARCHVALUES
+
+        public void Add(char value)
+        {
+            _builder.AddAsSortedSet(value);
+        }
+
+        internal CharacterSet Create(bool allowDestructive)
+        {
+            CharacterSet result;
+            if (allowDestructive)
+            {
+                result = new(_builder.Extract());
+            }
+            else
+            {
+                result = new(_builder.MakeArray());
+            }
+
+            return result;
+        }
+
+#else
+
         private char _mask = default;
-
-        public void AddRange(IEnumerable<char> values)
-        {
-            if (values is null) throw new ArgumentNullException(nameof(values));
-
-            foreach (var value in values)
-            {
-                Add(value);
-            }
-        }
-
-        public void AddRange(ReadOnlySpan<char> values)
-        {
-            foreach (var value in values)
-            {
-                Add(value);
-            }
-        }
 
         public void Add(char value)
         {
@@ -246,8 +321,6 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
                 _mask |= value;
             }
         }
-
-        public CharacterSet Create() => Create(allowDestructive: false);
 
         internal CharacterSet Create(bool allowDestructive)
         {
@@ -264,5 +337,31 @@ public readonly struct CharacterSet : IReadOnlyList<char>, IEquatable<CharacterS
 
             return result;
         }
+
+#endif
+
+        public void AddRange(IEnumerable<char> values)
+        {
+#if HAS_THROWNULL
+            ArgumentNullException.ThrowIfNull(values);
+#else
+            if (values is null) throw new ArgumentNullException(nameof(values));
+#endif
+
+            foreach (var value in values)
+            {
+                Add(value);
+            }
+        }
+
+        public void AddRange(ReadOnlySpan<char> values)
+        {
+            foreach (var value in values)
+            {
+                Add(value);
+            }
+        }
+
+        public CharacterSet Create() => Create(allowDestructive: false);
     }
 }
