@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Buffers;
-using System.Runtime.InteropServices;
 using System.Text;
 
 namespace WeCantSpell.Hunspell;
@@ -23,54 +22,56 @@ internal struct FlagParser
 
     public Encoding Encoding { get; set; }
 
-    public FlagValue ParseFlagOrDefault(ReadOnlySpan<char> text)
+    public readonly FlagValue ParseFlagOrDefault(ReadOnlySpan<char> text)
     {
         _ = TryParseFlag(text, out var result);
         return result;
     }
 
-    public bool TryParseFlag(ReadOnlySpan<char> text, out FlagValue value)
+    public readonly bool TryParseFlag(ReadOnlySpan<char> text, out FlagValue value)
     {
         return Mode switch
         {
             FlagParsingMode.Char => FlagValue.TryParseAsChar(text, out value),
-            FlagParsingMode.Uni => TryParseFlagAsUnicode(text, out value),
             FlagParsingMode.Long => FlagValue.TryParseAsLong(text, out value),
             FlagParsingMode.Num => FlagValue.TryParseAsNumber(text, out value),
-            _ => noOp(out value)
+            FlagParsingMode.Uni => TryParseFlagAsUnicode(text, out value),
+            _ => fail(out value)
         };
 
-        static bool noOp(out FlagValue value)
+        static bool fail(out FlagValue value)
         {
             value = default!;
             return false;
         }
     }
 
-    private bool TryParseFlagAsUnicode(ReadOnlySpan<char> text, out FlagValue value) =>
+    private readonly bool TryParseFlagAsUnicode(ReadOnlySpan<char> text, out FlagValue value) =>
         FlagValue.TryParseAsChar(ReDecodeConvertedStringAsUtf8(text, Encoding), out value);
 
-    public FlagValue[] ParseFlagsInOrder(ReadOnlySpan<char> text) => Mode switch
+    public readonly FlagValue[] ParseFlagsInOrder(ReadOnlySpan<char> text) => Mode switch
     {
         FlagParsingMode.Char => FlagValue.ParseAsChars(text),
-        FlagParsingMode.Uni => ParseFlagsInOrderAsUnicode(text),
         FlagParsingMode.Long => FlagValue.ParseAsLongs(text),
         FlagParsingMode.Num => FlagValue.ParseAsNumbers(text),
+        FlagParsingMode.Uni => ParseFlagsInOrderAsUnicode(text),
         _ => ThrowNotSupportedFlagMode<FlagValue[]>()
     };
 
-    private FlagValue[] ParseFlagsInOrderAsUnicode(ReadOnlySpan<char> text) => FlagValue.ParseAsChars(ReDecodeConvertedStringAsUtf8(text, Encoding));
+    private readonly FlagValue[] ParseFlagsInOrderAsUnicode(ReadOnlySpan<char> text) =>
+        FlagValue.ParseAsChars(ReDecodeConvertedStringAsUtf8(text, Encoding));
 
-    public FlagSet ParseFlagSet(ReadOnlySpan<char> text) => Mode switch
+    public readonly FlagSet ParseFlagSet(ReadOnlySpan<char> text) => Mode switch
     {
         FlagParsingMode.Char => FlagSet.ParseAsChars(text),
-        FlagParsingMode.Uni => ParseFlagSetAsUnicode(text),
         FlagParsingMode.Long => FlagSet.ParseAsLongs(text),
         FlagParsingMode.Num => FlagSet.ParseAsNumbers(text),
+        FlagParsingMode.Uni => ParseFlagSetAsUnicode(text),
         _ => ThrowNotSupportedFlagMode<FlagSet>()
     };
 
-    private FlagSet ParseFlagSetAsUnicode(ReadOnlySpan<char> text) => FlagSet.ParseAsChars(ReDecodeConvertedStringAsUtf8(text, Encoding));
+    private readonly FlagSet ParseFlagSetAsUnicode(ReadOnlySpan<char> text) =>
+        FlagSet.ParseAsChars(ReDecodeConvertedStringAsUtf8(text, Encoding));
 
     private static ReadOnlySpan<char> ReDecodeConvertedStringAsUtf8(ReadOnlySpan<char> decoded, Encoding encoding)
     {
@@ -86,7 +87,7 @@ internal struct FlagParser
 
         unsafe
         {
-            fixed (char* decodedPointer = &MemoryMarshal.GetReference(decoded))
+            fixed (char* decodedPointer = &System.Runtime.InteropServices.MemoryMarshal.GetReference(decoded))
             {
                 encodedBytes = new byte[Encoding.UTF8.GetByteCount(decodedPointer, decoded.Length)];
                 fixed (byte* encodedBytesPointer = &encodedBytes[0])
