@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
+﻿using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
 namespace WeCantSpell.Hunspell.Infrastructure;
@@ -8,18 +6,6 @@ namespace WeCantSpell.Hunspell.Infrastructure;
 static class CollectionsEx
 {
     internal const int CollectionPreallocationLimit = 16384;
-
-    public static bool Contains<T>(this T[] values, T value) => Array.IndexOf(values, value) >= 0;
-
-#if NO_DICTIONARY_GETVALUE
-
-    public static TValue? GetValueOrDefault<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dictionary, TKey key) =>
-        dictionary.TryGetValue(key, out var result) ? result : default;
-
-    public static TValue GetValueOrDefault<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dictionary, TKey key, TValue defaultValue) =>
-        dictionary.TryGetValue(key, out var result) ? result : defaultValue;
-
-#endif
 
     public static int RemoveDuplicates<T>(this List<T> list, IEqualityComparer<T> comparer)
     {
@@ -68,14 +54,6 @@ static class CollectionsEx
             }
 
             var duplicateCount = list.Count - writeIndex;
-
-#if DEBUG
-            if (duplicateCount < 0)
-            {
-                throw new InvalidOperationException();
-            }
-#endif
-
             if (duplicateCount > 0)
             {
                 list.RemoveRange(writeIndex, duplicateCount);
@@ -85,66 +63,11 @@ static class CollectionsEx
         }
     }
 
-    public static ImmutableArray<T> ToImmutable<T>(this ImmutableArray<T>.Builder builder, bool allowDestructive) =>
-        allowDestructive && builder.Capacity == builder.Count ? builder.MoveToImmutable() : builder.ToImmutable();
-
-    public static int RemoveSortedDuplicates<T>(ref T[] values) where T : notnull, IEquatable<T>
-    {
-        var shiftSpan = values.AsSpan();
-
-        var removed = 0;
-        while (shiftSpan.Length > 1)
-        {
-            if (shiftSpan[0].Equals(shiftSpan[1]))
-            {
-                for (var shiftIndex = 1; shiftIndex < shiftSpan.Length; shiftIndex++)
-                {
-                    shiftSpan[shiftIndex - 1] = shiftSpan[shiftIndex];
-                }
-
-                removed++;
-                shiftSpan = shiftSpan.Slice(0, shiftSpan.Length - 1);
-            }
-            else
-            {
-                shiftSpan = shiftSpan.Slice(1);
-            }
-        }
-
-        if (removed > 0)
-        {
-            Array.Resize(ref values, values.Length - removed);
-        }
-
-        return removed;
-    }
-
-    public static bool Contains(this List<string> list, ReadOnlySpan<char> value)
-    {
-        foreach (var item in list)
-        {
-            if (item is not null && value.Equals(item.AsSpan(), StringComparison.Ordinal))
-            {
-                return true;
-            }
-
-        }
-        return false;
-    }
-
-#if NO_KVP_DECONSTRUCT
-
-    public static void Deconstruct<TKey, TValue>(this KeyValuePair<TKey, TValue> pair, out TKey key, out TValue value)
-    {
-        key = pair.Key;
-        value = pair.Value;
-    }
-
-#endif
-
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static bool Any<T>(this List<T> list) => list.Count != 0;
 
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static bool Any<T>(this T[] array) => array.Length != 0;
+    public static void RemoveLast<T>(this List<T> list)
+    {
+        list.RemoveAt(list.Count - 1);
+    }
 }
