@@ -1,14 +1,18 @@
-﻿namespace WeCantSpell.Hunspell.Infrastructure;
+﻿using System.Runtime.CompilerServices;
+
+namespace WeCantSpell.Hunspell.Infrastructure;
 
 static class EnumEx
 {
-    public static bool HasFlag(this AffixConfigOptions value, AffixConfigOptions flag) => (value & flag) == flag;
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool HasFlagEx(this AffixConfigOptions value, AffixConfigOptions flag) => (value & flag) == flag;
 
-    public static bool AllowCross(this AffixEntryOptions value) => HasFlag(value, AffixEntryOptions.CrossProduct);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool HasFlagEx(this WordEntryOptions value, WordEntryOptions flag) => (value & flag) == flag;
 
-    public static bool HasFlag(this WordEntryOptions value, WordEntryOptions flag) => (value & flag) == flag;
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool HasFlagEx(this AffixEntryOptions value, AffixEntryOptions flag) => (value & flag) == flag;
 
-    public static bool HasFlag(this AffixEntryOptions value, AffixEntryOptions flag) => (value & flag) == flag;
-
-    public static bool HasFlag(this SpellCheckResultType value, SpellCheckResultType flag) => (value & flag) == flag;
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static bool HasFlagEx(this SpellCheckResultType value, SpellCheckResultType flag) => (value & flag) == flag;
 }
