@@ -6,6 +6,8 @@ namespace WeCantSpell.Hunspell.Infrastructure;
 
 static class HunspellTextFunctions
 {
+
+#if false // This isn't used anymore but I want to keep it around as it was tricky to port
     public static bool IsReverseSubset(string s1, ReadOnlySpan<char> s2)
     {
         return s1.Length <= s2.Length && check(s1.AsSpan(), s2);
@@ -27,6 +29,7 @@ static class HunspellTextFunctions
             return true;
         }
     }
+#endif
 
     public static bool IsSubset(string s1, ReadOnlySpan<char> s2)
     {
@@ -51,9 +54,8 @@ static class HunspellTextFunctions
     public static bool IsNumericWord(ReadOnlySpan<char> word)
     {
         byte state = 0; // 0 = begin, 1 = number, 2 = separator
-        for (var i = 0; i < word.Length; i++)
+        foreach (var c in word)
         {
-            var c = word[i];
             if (char.IsNumber(c))
             {
                 state = 1;
@@ -94,28 +96,42 @@ static class HunspellTextFunctions
 
     public static int CountMatchingFromRight(string text, char character)
     {
-        var lastIndex = text.Length - 1;
-        var searchIndex = lastIndex;
+        var searchIndex = text.Length - 1;
         for (; searchIndex >= 0 && text[searchIndex] == character; searchIndex--) ;
 
-        return lastIndex - searchIndex;
+        return text.Length - searchIndex - 1;
     }
 
     public static int CountMatchingFromRight(ReadOnlySpan<char> text, char character)
     {
-        var lastIndex = text.Length - 1;
-        var searchIndex = lastIndex;
+        var searchIndex = text.Length - 1;
         for (; searchIndex >= 0 && text[searchIndex] == character; searchIndex--) ;
 
-        return lastIndex - searchIndex;
+        return text.Length - searchIndex - 1;
+    }
+
+    public static int CountMatchesFromLeft<T>(this ReadOnlySpan<T> a, ReadOnlySpan<T> b) where T : notnull, IEquatable<T>
+    {
+        var minLength = Math.Min(a.Length, b.Length);
+        var count = 0;
+        for (; count < minLength && a[count].Equals(b[count]); count++) ;
+        return count;
+    }
+
+    public static int CountMatchesFromRight<T>(this ReadOnlySpan<T> a, ReadOnlySpan<T> b) where T : notnull, IEquatable<T>
+    {
+        var minLength = Math.Min(a.Length, b.Length);
+        var count = 0;
+        for (; count < minLength && a[a.Length - 1 - count].Equals(b[b.Length - 1 - count]); count++) ;
+        return count;
     }
 
     /// <summary>
-    /// This is a character class function used within Hunspell to determine if a character is an ASCII letter.
+    /// This is a character class function used within Hunspell to determine if a character is an ASCII letter or something else.
     /// </summary>
     /// <param name="ch">The character value to check.</param>
-    /// <returns><c>true</c> is a given character is an ASCII letter.</returns>
-    public static bool MyIsAlpha(char ch) => ch >= 128 || char.IsLetter(ch);
+    /// <returns><c>true</c> is a given character is an ASCII letter or something else.</returns>
+    public static bool MyIsAlpha(char ch) => ch is (>= 'A' and <= 'Z') or (>= 'a' and <= 'z') or >= (char)128;
 
     public static string MakeInitCap(string s, TextInfo textInfo)
     {
@@ -125,7 +141,7 @@ static class HunspellTextFunctions
             var expectedFirstLetter = textInfo.ToUpper(actualFirstLetter);
             if (expectedFirstLetter != actualFirstLetter)
             {
-                return ReplaceFirstLetter(expectedFirstLetter, s.AsSpan());
+                return StringEx.ConcatString(expectedFirstLetter, s.AsSpan(1));
             }
         }
 
@@ -140,24 +156,11 @@ static class HunspellTextFunctions
             var expectedFirstLetter = textInfo.ToUpper(actualFirstLetter);
             if (expectedFirstLetter != actualFirstLetter)
             {
-                return ReplaceFirstLetter(expectedFirstLetter, s).AsSpan();
+                return StringEx.ConcatString(expectedFirstLetter, s.Slice(1)).AsSpan();
             }
         }
 
         return s;
-    }
-
-    private static string ReplaceFirstLetter(char firstLetter, ReadOnlySpan<char> baseText)
-    {
-        var builder = StringBuilderPool.Get(baseText.Length);
-        builder.Append(firstLetter);
-
-        if (baseText.Length > 1)
-        {
-            builder.Append(baseText.Slice(1));
-        }
-
-        return StringBuilderPool.GetStringAndReturn(builder);
     }
 
     /// <summary>
@@ -173,7 +176,7 @@ static class HunspellTextFunctions
             var expectedFirstLetter = textInfo.ToLower(actualFirstLetter);
             if (expectedFirstLetter != actualFirstLetter)
             {
-                return ReplaceFirstLetter(expectedFirstLetter, s.AsSpan());
+                return StringEx.ConcatString(expectedFirstLetter, s.AsSpan(1));
             }
         }
 
@@ -265,6 +268,8 @@ static class HunspellTextFunctions
             return CapitalizationType.Huh;
         }
 
-        static bool charIsNotNeutral(char c, TextInfo textInfo) => (c < 127 || textInfo.ToUpper(c) != c) && char.IsLower(c);
+        static bool charIsNotNeutral(char c, TextInfo textInfo) => c < 128
+            ? c is >= 'a' and <= 'z' // For ASCII, only the a-z range needs to be checked
+            : (char.IsLower(c) && textInfo.ToUpper(c) != c); // Outside ASCII, use the framework combined with the uppercase thing
     }
 }
