@@ -3,7 +3,7 @@ using System.Threading;
 
 namespace WeCantSpell.Hunspell;
 
-ref struct OperationTimedLimiter
+struct OperationTimedLimiter
 {
     public OperationTimedLimiter(TimeSpan timeLimit, CancellationToken cancellationToken)
     {
@@ -32,7 +32,7 @@ ref struct OperationTimedLimiter
     }
 }
 
-ref struct OperationTimedCountLimiter
+struct OperationTimedCountLimiter
 {
     /// <summary>
     /// This is the number of operations that are added to a limiter if it runs out of operations before the time limit has expired.
@@ -41,10 +41,6 @@ ref struct OperationTimedCountLimiter
 
     public OperationTimedCountLimiter(TimeSpan timeLimit, int countLimit, CancellationToken cancellationToken)
     {
-#if DEBUG
-        if (countLimit < 0) throw new ArgumentOutOfRangeException(nameof(countLimit));
-#endif
-
         _timer = new ExpirationTimer(timeLimit);
         _cancellationToken = cancellationToken;
         _counter = countLimit;
