@@ -16,33 +16,27 @@ public readonly struct WordEntryDetail : IEquatable<WordEntryDetail>
 
     public WordEntryDetail()
     {
-        Flags = FlagSet.Empty;
         Morphs = MorphSet.Empty;
+        Flags = FlagSet.Empty;
         Options = WordEntryOptions.None;
     }
 
     public WordEntryDetail(FlagSet flags, MorphSet morphs, WordEntryOptions options)
     {
-        Flags = flags;
         Morphs = morphs;
+        Flags = flags;
         Options = options;
     }
 
-    public FlagSet Flags { get; }
-
     public MorphSet Morphs { get; }
+
+    public FlagSet Flags { get; }
 
     public WordEntryOptions Options { get; }
 
-    public bool HasFlags => Flags.HasItems;
-
     public bool ContainsFlag(FlagValue flag) => Flags.Contains(flag);
 
-    public bool ContainsAnyFlags(FlagValue a, FlagValue b) => Flags.ContainsAny(a, b);
-
-    public bool ContainsAnyFlags(FlagValue a, FlagValue b, FlagValue c) => Flags.ContainsAny(a, b, c);
-
-    public bool ContainsAnyFlags(FlagValue a, FlagValue b, FlagValue c, FlagValue d) => Flags.ContainsAny(a, b, c, d);
+    public bool ContainsAnyFlags(FlagSet set) => Flags.ContainsAny(set);
 
     public bool Equals(WordEntryDetail other) =>
         other.Options == Options
