@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Buffers;
-using System.Collections.Generic;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using System.Threading;
@@ -16,7 +15,7 @@ public partial class WordList
         internal const string DefaultXmlToken = "<?xml?>";
         internal const string DefaultXmlTokenCheckPrefix = "<?xml";
 
-        private static QueryOptions DefaultOptions { get; } = new();
+        private static readonly QueryOptions DefaultOptions = new();
 
         internal Query(WordList wordList, QueryOptions? options, CancellationToken cancellationToken)
         {
@@ -25,13 +24,13 @@ public partial class WordList
             TextInfo = Affix.Culture.TextInfo;
             Options = options ?? DefaultOptions;
             CancellationToken = cancellationToken;
-            SpellCandidateStack = new();
-            Prefix = null;
+            _spellCandidateStack = new();
+            _prefix = null;
             //PrefixAppend = null;
-            Suffix = null;
-            SuffixFlag = default;
-            SuffixExtra = false;
-            SuffixAppend = null;
+            _suffix = null;
+            _suffixFlag = default;
+            _suffixExtra = false;
+            _suffixAppend = null;
         }
 
         internal Query(in Query source)
@@ -41,22 +40,22 @@ public partial class WordList
             TextInfo = Affix.Culture.TextInfo;
             Options = source.Options;
             CancellationToken = source.CancellationToken;
-            SpellCandidateStack = source.SpellCandidateStack;
-            Prefix = null;
+            _spellCandidateStack = source._spellCandidateStack;
+            _prefix = null;
             //PrefixAppend = null;
-            Suffix = null;
-            SuffixFlag = default;
-            SuffixExtra = false;
-            SuffixAppend = null;
+            _suffix = null;
+            _suffixFlag = default;
+            _suffixExtra = false;
+            _suffixAppend = null;
         }
 
-        public WordList WordList { get; }
+        public readonly WordList WordList;
 
-        public AffixConfig Affix { get; }
+        public readonly AffixConfig Affix;
 
-        public QueryOptions Options { get; }
+        public readonly QueryOptions Options;
 
-        public TextInfo TextInfo { get; }
+        public readonly TextInfo TextInfo;
 
         /// <summary>
         /// A cancellation token that can be used to request the termination of a check or suggest operation.
@@ -68,114 +67,95 @@ public partial class WordList
         /// of results. Even without cancellation, the consistency of results can't be gauranteed
         /// due to the use of timing checks throughout the code.
         /// </remarks>
-        public CancellationToken CancellationToken { get; }
+        public readonly CancellationToken CancellationToken;
 
-        public CandidateStack SpellCandidateStack { get; private set; }
+        internal CandidateStack _spellCandidateStack;
 
-        private PrefixEntry? Prefix { get; set; }
+        private PrefixEntry? _prefix;
 
         //private string? PrefixAppend { get; set; } // Previous prefix for counting syllables of the prefix.
 
-        private SuffixEntry? Suffix { get; set; }
+        private SuffixEntry? _suffix;
 
-        private FlagValue SuffixFlag { get; set; }
+        private FlagValue _suffixFlag;
 
         /// <summary>
-        /// Modifier for syllable count of <see cref="SuffixAppend"/>.
+        /// Modifier for syllable count of <see cref="_suffixAppend"/>.
         /// </summary>
-        private bool SuffixExtra { get; set; }
+        private bool _suffixExtra;
 
         /// <summary>
         /// Previous suffix for counting syllables of the suffix.
         /// </summary>
-        private string? SuffixAppend { get; set; }
+        private string? _suffixAppend;
 
         private void ClearPrefix()
         {
-            Prefix = null;
+            _prefix = null;
         }
 
         private void ClearSuffix()
         {
-            Suffix = null;
+            _suffix = null;
         }
 
         private void ClearSuffixAndFlag()
         {
             ClearSuffix();
-            SuffixFlag = default;
+            _suffixFlag = default;
         }
 
         private void ClearSuffixAppendAndExtra()
         {
-            SuffixAppend = null;
-            SuffixExtra = false;
+            _suffixAppend = null;
+            _suffixExtra = false;
         }
 
         private void ClearAllAppendAndExtra()
         {
             //PrefixAppend = null;
-            SuffixAppend = null;
-            SuffixExtra = false;
+            _suffixAppend = null;
+            _suffixExtra = false;
         }
 
         private void SetPrefix(PrefixEntry prefix)
         {
-            Prefix = prefix;
+            _prefix = prefix;
         }
 
         private void SetSuffix(SuffixEntry suffix)
         {
-            Suffix = suffix;
+            _suffix = suffix;
         }
 
         private void SetSuffixFlag(FlagValue flag)
         {
-            SuffixFlag = flag;
+            _suffixFlag = flag;
         }
 
         private void SetSuffixExtra(bool extra)
         {
-            SuffixExtra = extra;
+            _suffixExtra = extra;
         }
 
         private void SetSuffixAppend(string append)
         {
-            SuffixAppend = append;
+            _suffixAppend = append;
         }
 
-        private bool AffixContainsContClass(FlagValue value) =>
+        private readonly bool AffixContainsContClass(FlagValue value) =>
             value.HasValue
             &&
-            (
-                (Prefix is not null && Prefix.ContainsContClass(value))
-                ||
-                (Suffix?.ContainsContClass(value)).GetValueOrDefault()
-            );
+            ((_prefix?.ContainsContClass(value)).GetValueOrDefault() || (_suffix?.ContainsContClass(value)).GetValueOrDefault());
 
-        private bool ContainFlagsOrBlockSuggest(in WordEntryDetail rv, bool isSug, FlagValue a, FlagValue b) =>
-            rv.HasFlags
-            &&
-            (
-                rv.ContainsAnyFlags(a, b)
-                ||
-                (isSug && rv.ContainsFlag(Affix.NoSuggest))
-            );
-
-        private bool ContainFlagsOrBlockSuggest(in WordEntryDetail rv, bool isSug, FlagValue a, FlagValue b, FlagValue c) =>
-            rv.HasFlags
-            &&
-            (
-                rv.ContainsAnyFlags(a, b, c)
-                ||
-                (isSug && rv.ContainsFlag(Affix.NoSuggest))
-            );
+        private readonly bool AffixContainsAnyContClass(FlagSet values) =>
+            (_prefix?.ContainsAnyContClass(values)).GetValueOrDefault() || (_suffix?.ContainsAnyContClass(values)).GetValueOrDefault();
 
         public WordEntry? CheckWord(string word, ref SpellCheckResultType info, out string? root)
         {
             root = null;
 
-            if (string.IsNullOrEmpty(word))
+            if (word.Length == 0)
             {
                 return null;
             }
@@ -197,67 +177,23 @@ public partial class WordList
                 word = word.GetReversed();
             }
 
+
             // look word in hash table
-            var details = WordList.FindEntryDetailsByRootWord(word);
-            if (details.Length != 0)
+            if (!TryCheckWord_HashTable(word, ref info, out var he))
             {
-                var heDetails = details[0];
+                // check with affixes
 
-                // check forbidden and onlyincompound words
-                if (heDetails.ContainsFlag(Affix.ForbiddenWord))
-                {
-                    info |= SpellCheckResultType.Forbidden;
-
-                    if (heDetails.ContainsFlag(Affix.CompoundFlag) && Affix.IsHungarian)
-                    {
-                        info |= SpellCheckResultType.Compound;
-                    }
-
-                    return null;
-                }
-
-                // he = next not needaffix, onlyincompound homonym or onlyupcase word
-                var heIndex = 0;
-                while (
-                    heDetails.HasFlags
-                    &&
-                    (
-                        heDetails.ContainsAnyFlags(Affix.NeedAffix, Affix.OnlyInCompound)
-                        ||
-                        hasSpecialInitCap(info, heDetails)
-                    )
-                )
-                {
-                    heIndex++;
-                    if (heIndex < details.Length)
-                    {
-                        heDetails = details[heIndex];
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-
-                if (heIndex < details.Length)
-                {
-                    return new WordEntry(word, heDetails);
-                }
-            }
-
-            // check with affixes
-            {
                 // try stripping off affixes
-                var he = AffixCheck(word.AsSpan(), default, CompoundOptions.Not);
+                he = AffixCheck(word.AsSpan(), default, CompoundOptions.Not);
 
                 // check compound restriction and onlyupcase
                 if (
                     he is not null
                     &&
                     (
-                        he.ContainsFlag(Affix.OnlyInCompound)
-                        ||
-                        hasSpecialInitCap(info, he.Detail)
+                        info.HasFlagEx(SpellCheckResultType.InitCap)
+                            ? he.ContainsAnyFlags(Affix.Flags_OnlyInCompound_OnlyUpcase)
+                            : he.ContainsFlag(Affix.OnlyInCompound)
                     )
                 )
                 {
@@ -282,43 +218,7 @@ public partial class WordList
                 else if (Affix.HasCompound)
                 {
                     // try check compound word
-                    var rwords = new IncrementalWordList();
-
-                    // first allow only 2 words in the compound
-                    var setinfo = SpellCheckResultType.Compound2 | info;
-                    he = CompoundCheck(word.AsSpan(), 0, 0, 100, null, rwords, huMovRule: false, isSug: false, ref setinfo);
-                    info = setinfo & ~SpellCheckResultType.Compound2; // unset Compound2
-
-                    // if not 2-word compoud word, try with 3 or more words
-                    // (only if original info didn't forbid it)
-                    if (he is null && !info.HasFlag(SpellCheckResultType.Compound2))
-                    {
-                        info &= ~SpellCheckResultType.Compound2;
-                        he = CompoundCheck(word.AsSpan(), 0, 0, 100, null, rwords, huMovRule: false, isSug: false, ref info);
-                        // accept the compound with 3 or more words only if it is
-                        // - not a dictionary word with a typo and
-                        // - not two words written separately,
-                        // - or if it's an arbitrary number accepted by compound rules (e.g. 999%)
-                        if (he is not null && word.Length > 0 && !char.IsDigit(word[0]))
-                        {
-                            var querySuggest = new QuerySuggest(WordList, Options, CancellationToken)
-                            {
-                                TestSimpleSuggestion = true
-                            };
-
-                            bool onlyCompoundSug = false;
-                            if (querySuggest.Suggest(new List<string>(), word, ref onlyCompoundSug))
-                            {
-                                he = null;
-                            }
-                        }
-                    }
-
-                    if (he is null && word.EndsWith('-') && Affix.IsHungarian)
-                    {
-                        // LANG_hu section: `moving rule' with last dash
-                        he = CompoundCheck(word.AsSpan(0, word.Length - 1), -5, 0, 100, null, rwords, huMovRule: true, isSug: false, ref info);
-                    }
+                    TryCheckWord_Compound(word, ref info, out he);
 
                     if (he is not null)
                     {
@@ -331,95 +231,109 @@ public partial class WordList
                         info |= SpellCheckResultType.Compound;
                     }
                 }
-
-                return he;
             }
 
-            static bool hasSpecialInitCap(SpellCheckResultType info, in WordEntryDetail he) =>
-                EnumEx.HasFlag(info, SpellCheckResultType.InitCap) && he.ContainsFlag(SpecialFlags.OnlyUpcaseFlag);
+            return he;
         }
 
-        private WordEntryDetail? CompoundCheckWordSearch_OnlyCpdRule(
-            WordEntryDetail[] searchEntryDetails,
-            FlagValue scpdPatternEntryCondition,
-            ref IncrementalWordList? words,
-            IncrementalWordList rwords)
+        /// <summary>
+        /// look word in hash table
+        /// </summary>
+        /// <returns><c>true</c> when result should be used</returns>
+        private readonly bool TryCheckWord_HashTable(string word, ref SpellCheckResultType info, out WordEntry? result)
         {
-            foreach (var searchEntryDetail in searchEntryDetails)
+            result = null;
+            var details = WordList.FindEntryDetailsByRootWord(word);
+            if (details.Length == 0)
             {
-                if (!searchEntryDetail.ContainsFlag(Affix.NeedAffix))
+                return false;
+            }
+
+            ref readonly var heDetails = ref details[0];
+
+            // check forbidden and onlyincompound words
+            if (heDetails.ContainsFlag(Affix.ForbiddenWord))
+            {
+                info |= SpellCheckResultType.Forbidden;
+
+                if (Affix.IsHungarian && heDetails.ContainsFlag(Affix.CompoundFlag))
                 {
-                    // NOTE: do not reorder DefCompoundCheck calls or other conditions due to side effects
+                    info |= SpellCheckResultType.Compound;
+                }
 
-                    bool defCpdCheck;
-                    if (words is not null)
-                    {
-                        defCpdCheck = DefCompoundCheck(words, searchEntryDetail, false);
-                    }
-                    else if (DefCompoundCheck(rwords, searchEntryDetail, false))
-                    {
-                        words = rwords;
-                        defCpdCheck = true;
-                    }
-                    else
-                    {
-                        defCpdCheck = false;
-                    }
+                return true;
+            }
 
-                    // NOTE: do not reorder DefCompoundCheck calls or other conditions due to side effects
+            var heFlags = info.HasFlagEx(SpellCheckResultType.InitCap) ? Affix.Flags_NeedAffix_OnlyInCompound_OnlyUpcase : Affix.Flags_NeedAffix_OnlyInCompound;
 
-                    if (defCpdCheck && (scpdPatternEntryCondition.IsZero || searchEntryDetail.ContainsFlag(scpdPatternEntryCondition)))
+            // he = next not needaffix, onlyincompound homonym or onlyupcase word
+            var heIndex = 0;
+            while (heDetails.ContainsAnyFlags(heFlags))
+            {
+                heIndex++;
+                if (heIndex < details.Length)
+                {
+                    heDetails = ref details[heIndex];
+                }
+                else
+                {
+                    break;
+                }
+            }
+
+            if (heIndex < details.Length)
+            {
+                result = new WordEntry(word, heDetails);
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// try check compound word
+        /// </summary>
+        private void TryCheckWord_Compound(string word, ref SpellCheckResultType info, out WordEntry? he)
+        {
+            // try check compound word
+            var rwords = IncrementalWordList.GetRoot();
+
+            // first allow only 2 words in the compound
+            var setinfo = SpellCheckResultType.Compound2 | info;
+            he = CompoundCheck(word.AsSpan(), 0, 0, 100, rwords, huMovRule: false, isSug: false, setinfo);
+            info = setinfo & ~SpellCheckResultType.Compound2; // unset Compound2
+
+            // if not 2-word compoud word, try with 3 or more words
+            // (only if original info didn't forbid it)
+            if (he is null && !info.HasFlagEx(SpellCheckResultType.Compound2))
+            {
+                info &= ~SpellCheckResultType.Compound2;
+                he = CompoundCheck(word.AsSpan(), 0, 0, 100, rwords, huMovRule: false, isSug: false, info);
+                // accept the compound with 3 or more words only if it is
+                // - not a dictionary word with a typo and
+                // - not two words written separately,
+                // - or if it's an arbitrary number accepted by compound rules (e.g. 999%)
+                if (he is not null && word.Length > 0 && !char.IsDigit(word[0]))
+                {
+                    var querySuggest = new QuerySuggest(WordList, Options, CancellationToken, testSimpleSuggestion: true);
+                    var onlyCompoundSug = false;
+                    if (querySuggest.Suggest([], word, ref onlyCompoundSug))
                     {
-                        return searchEntryDetail;
+                        he = null;
                     }
                 }
             }
 
-            return null;
-        }
-
-        private WordEntryDetail? CompoundCheckWordSearch_NormalNoWordList(
-            WordEntryDetail[] searchEntryDetails,
-            FlagValue scpdPatternEntryCondition,
-            FlagValue compoundPart)
-        {
-            foreach (var searchEntryDetail in searchEntryDetails)
+            if (he is null && word.EndsWith('-') && Affix.IsHungarian)
             {
-                if (
-                    !searchEntryDetail.ContainsFlag(Affix.NeedAffix)
-                    &&
-                    searchEntryDetail.ContainsAnyFlags(Affix.CompoundFlag, scpdPatternEntryCondition, compoundPart)
-                )
-                {
-                    return searchEntryDetail;
-                }
+                // LANG_hu section: `moving rule' with last dash
+                he = CompoundCheck(word.AsSpan(0, word.Length - 1), -5, 0, 100, rwords, huMovRule: true, isSug: false, info);
             }
 
-            return null;
+            IncrementalWordList.ReturnRoot(ref rwords);
         }
 
-        private WordEntryDetail? CompoundCheckWordSearch_NormalWithExistingWordList(
-            WordEntryDetail[] searchEntryDetails,
-            FlagValue scpdPatternEntryCondition)
-        {
-            foreach (var searchEntryDetail in searchEntryDetails)
-            {
-                if (
-                    !searchEntryDetail.ContainsFlag(Affix.NeedAffix)
-                    &&
-                    searchEntryDetail.ContainsFlag(Affix.CompoundBegin)
-                    &&
-                    (scpdPatternEntryCondition.IsZero || searchEntryDetail.ContainsFlag(scpdPatternEntryCondition))
-                )
-                {
-                    return searchEntryDetail;
-                }
-            }
-
-            return null;
-        }
-
-        private WordEntry? HomonymWordSearch(ReadOnlySpan<char> homonymWord, IncrementalWordList? words, FlagValue condition2, bool scpdIsZero)
+        private readonly WordEntry? HomonymWordSearch(ReadOnlySpan<char> homonymWord, IncrementalWordList? words, FlagValue condition2, bool scpdIsZero)
         {
             // perhaps without prefix
             if (TryLookupDetails(homonymWord, out var homonymWordString, out var details) && details.Length > 0)
@@ -432,15 +346,8 @@ public partial class WordList
                         &&
                         (
                             words is null
-                            ? (
-                                homonymCandidate.ContainsFlag(Affix.CompoundFlag)
-                                ||
-                                homonymCandidate.ContainsFlag(Affix.CompoundEnd)
-                            )
-                            : (
-                                Affix.CompoundRules.HasItems
-                                && DefCompoundCheck(words.CreateIncremented(), homonymCandidate, true)
-                            )
+                            ? homonymCandidate.ContainsAnyFlags(Affix.Flags_CompoundFlag_CompoundEnd)
+                            : (Affix.CompoundRules.HasItems && DefCompoundCheck(words.CreateIncremented(), homonymCandidate, true))
                         )
                         &&
                         (
@@ -458,15 +365,15 @@ public partial class WordList
             return null;
         }
 
-        public WordEntry? CompoundCheck(ReadOnlySpan<char> word, int wordNum, int numSyllable, int maxwordnum, IncrementalWordList? words, IncrementalWordList rwords, bool huMovRule, bool isSug, ref SpellCheckResultType info)
+        public WordEntry? CompoundCheck(ReadOnlySpan<char> word, int wordNum, int numSyllable, int maxwordnum, IncrementalWordList rwords, bool huMovRule, bool isSug, SpellCheckResultType info)
         {
             // add a time limit to handle possible
             // combinatorical explosion of the overlapping words
             var opLimiter = new OperationTimedLimiter(Options.TimeLimitCompoundCheck, CancellationToken);
-            return CompoundCheck(word, wordNum, numSyllable, maxwordnum, words, rwords, huMovRule, isSug, ref info, ref opLimiter);
+            return CompoundCheck(word, wordNum, numSyllable, maxwordnum, null, rwords, huMovRule, isSug, info, ref opLimiter);
         }
 
-        public WordEntry? CompoundCheck(ReadOnlySpan<char> word, int wordNum, int numSyllable, int maxwordnum, IncrementalWordList? words, IncrementalWordList rwords, bool huMovRule, bool isSug, ref SpellCheckResultType info, ref OperationTimedLimiter opLimiter)
+        public WordEntry? CompoundCheck(ReadOnlySpan<char> word, int wordNum, int numSyllable, int maxwordnum, IncrementalWordList? words, IncrementalWordList rwords, bool huMovRule, bool isSug, SpellCheckResultType info, ref OperationTimedLimiter opLimiter)
         {
             int oldnumsyllable, oldnumsyllable2, oldwordnum, oldwordnum2;
             WordEntry? rv;
@@ -492,10 +399,6 @@ public partial class WordList
             var cmax = word.Length - cmin + 1;
 
             var st = new SimulatedCString(word);
-
-            var conditionBypassAllCompounds = Affix.CompoundFlag.IsZero && Affix.CompoundBegin.IsZero && Affix.CompoundMiddle.IsZero;
-
-            var movCompoundOptions = huMovRule ? CompoundOptions.Other : CompoundOptions.Begin;
 
             for (var i = cmin; i < cmax; i++)
             {
@@ -603,49 +506,48 @@ public partial class WordList
                                         continue;
                                     }
 
-                                    WordEntryDetail? rvDetail;
-
                                     if (onlycpdrule)
                                     {
                                         if (Affix.CompoundRules.HasItems && (wordNum == 0 || words is not null))
                                         {
-                                            rvDetail = CompoundCheckWordSearch_OnlyCpdRule(
+                                            rv = CompoundCheckWordSearch_OnlyCpdRule(
                                                 searchEntryDetails,
                                                 scpdPatternEntryCondition,
                                                 ref words,
-                                                rwords);
+                                                rwords)
+                                                ?.ToEntry(searchEntryWord);
                                         }
                                         else
                                         {
-                                            rvDetail = null;
+                                            rv = null;
                                         }
                                     }
-                                    else if (conditionBypassAllCompounds)
+                                    else if (Affix.CompoundFlag.IsZero && Affix.CompoundBegin.IsZero && Affix.CompoundMiddle.IsZero)
                                     {
-                                        rvDetail = null;
+                                        rv = null;
                                     }
                                     else
                                     {
                                         if (words is null)
                                         {
-                                            rvDetail = CompoundCheckWordSearch_NormalNoWordList(
+                                            rv = CompoundCheckWordSearch_NormalNoWordList(
                                                 searchEntryDetails,
                                                 scpdPatternEntryCondition,
-                                                wordNum == 0 ? Affix.CompoundBegin : Affix.CompoundMiddle);
+                                                wordNum == 0 ? Affix.CompoundBegin : Affix.CompoundMiddle)
+                                                ?.ToEntry(searchEntryWord);
                                         }
                                         else if (wordNum == 0)
                                         {
-                                            rvDetail = CompoundCheckWordSearch_NormalWithExistingWordList(
+                                            rv = CompoundCheckWordSearch_NormalWithExistingWordList(
                                                 searchEntryDetails,
-                                                scpdPatternEntryCondition);
+                                                scpdPatternEntryCondition)
+                                                ?.ToEntry(searchEntryWord);
                                         }
                                         else
                                         {
-                                            rvDetail = null;
+                                            rv = null;
                                         }
                                     }
-
-                                    rv = rvDetail?.ToEntry(searchEntryWord);
                                 }
                                 else
                                 {
@@ -667,26 +569,7 @@ public partial class WordList
 
                             if (Affix.CompoundFlag.HasValue)
                             {
-                                rv = PrefixCheck(st.TerminatedSpan, movCompoundOptions, Affix.CompoundFlag);
-                                if (rv is null)
-                                {
-                                    rv = SuffixCheck(st.TerminatedSpan, AffixEntryOptions.None, null, new FlagValue(), Affix.CompoundFlag, movCompoundOptions);
-                                    if (rv is null && Affix.CompoundMoreSuffixes)
-                                    {
-                                        rv = SuffixCheckTwoSfx(st.TerminatedSpan, AffixEntryOptions.None, null, Affix.CompoundFlag);
-                                    }
-
-                                    if (
-                                        rv is not null
-                                        &&
-                                        !huMovRule
-                                        &&
-                                        (Suffix?.ContainsAnyContClass(Affix.CompoundForbidFlag, Affix.CompoundEnd)).GetValueOrDefault()
-                                    )
-                                    {
-                                        rv = null;
-                                    }
-                                }
+                                rv = CompoundCheckWordSearch_AffixCheckCompoundFlag(st.TerminatedSpan, huMovRule);
                             }
 
                             if (rv is not null)
@@ -695,68 +578,21 @@ public partial class WordList
                             }
                             else if (wordNum == 0 && Affix.CompoundBegin.HasValue)
                             {
-                                rv = SuffixCheck(st.TerminatedSpan, AffixEntryOptions.None, null, default, Affix.CompoundBegin, movCompoundOptions);
-
-                                if(rv is null)
-                                {
-                                    if(Affix.CompoundMoreSuffixes)
-                                    {
-                                        rv = SuffixCheckTwoSfx(st.TerminatedSpan, AffixEntryOptions.None, null, Affix.CompoundBegin);
-                                        if (rv is not null)
-                                        {
-                                            checkedPrefix = true;
-                                        }
-                                    }
-
-                                    if (rv is null)
-                                    {
-                                        rv = PrefixCheck(st.TerminatedSpan, movCompoundOptions, Affix.CompoundBegin);
-                                        if (rv is not null)
-                                        {
-                                            checkedPrefix = true;
-                                        }
-                                    }
-                                }
-                                else
-                                {
-                                    checkedPrefix = true;
-                                }
+                                rv = CompoundCheckWordSearch_AffixCheckCompoundBegin(st.TerminatedSpan, huMovRule);
+                                checkedPrefix |= rv is not null;
                             }
                             else if (wordNum > 0 && Affix.CompoundMiddle.HasValue)
                             {
-                                rv = SuffixCheck(st.TerminatedSpan, AffixEntryOptions.None, null, default, Affix.CompoundMiddle, movCompoundOptions);
-                                if (rv is null)
-                                {
-                                    if (Affix.CompoundMoreSuffixes)
-                                    {
-                                        rv = SuffixCheckTwoSfx(st.TerminatedSpan, AffixEntryOptions.None, null, Affix.CompoundMiddle);
-                                        if (rv is not null)
-                                        {
-                                            checkedPrefix = true;
-                                        }
-                                    }
-
-                                    if (rv is null)
-                                    {
-                                        rv = PrefixCheck(st.TerminatedSpan, movCompoundOptions, Affix.CompoundMiddle);
-                                        if (rv is not null)
-                                        {
-                                            checkedPrefix = true;
-                                        }
-                                    }
-                                }
-                                else
-                                {
-                                    checkedPrefix = true;
-                                }
+                                rv = CompoundCheckWordSearch_AffixCheckCompoundMiddle(st.TerminatedSpan, huMovRule);
+                                checkedPrefix |= rv is not null;
                             }
-
                         }
                         else
                         {
                             affixed = false;
-
-                            if (ContainFlagsOrBlockSuggest(rv.Detail, isSug, Affix.ForbiddenWord, Affix.NeedAffix, SpecialFlags.OnlyUpcaseFlag))
+                            if (
+                                rv.ContainsAnyFlags(isSug ? Affix.Flags_NeedAffix_ForbiddenWord_OnlyUpcase_NoSuggest : Affix.Flags_NeedAffix_ForbiddenWord_OnlyUpcase)
+                            )
                             {
                                 // else check forbiddenwords and needaffix
                                 st[i] = ch;
@@ -765,38 +601,29 @@ public partial class WordList
                             }
                         }
 
-                        if(
-                            rv is not null
-                            &&
-                            !huMovRule
-                            &&
-                            (
-                                // check non_compound flag in suffix and prefix
-                                AffixContainsContClass(Affix.CompoundForbidFlag)
-                                ||
-                                (
-                                    !checkedPrefix
-                                    &&
-                                    (
-                                        // check compoundend flag in suffix and prefix
-                                        AffixContainsContClass(Affix.CompoundEnd)
-                                        ||
-                                        // check compoundmiddle flag in suffix and prefix
-                                        (wordNum == 0 && AffixContainsContClass(Affix.CompoundMiddle))
-                                    )
-                                )
-                            )
-                        )
+                        if (rv is not null && !huMovRule)
                         {
-                            rv = null;
+                            if (
+                                checkedPrefix
+                                    ? AffixContainsContClass(Affix.CompoundForbidFlag) // check non_compound flag in suffix and prefix
+                                    : wordNum == 0
+                                        // check non_compound flag in suffix and prefix
+                                        // check compoundend flag in suffix and prefix
+                                        // check compoundmiddle flag in suffix and prefix
+                                        ? AffixContainsAnyContClass(Affix.Flags_CompoundForbid_CompoundMiddle_CompoundEnd)
+                                        // check non_compound flag in suffix and prefix
+                                        // check compoundend flag in suffix and prefix
+                                        : AffixContainsAnyContClass(Affix.Flags_CompoundForbid_CompoundEnd)
+                            )
+                            {
+                                rv = null;
+                            }
                         }
-
-                        bool firstWordCompoundAcceptable;
 
                         if (rv is not null)
                         {
                             // check forbiddenwords
-                            if (ContainFlagsOrBlockSuggest(rv.Detail, isSug, Affix.ForbiddenWord, SpecialFlags.OnlyUpcaseFlag))
+                            if (rv.ContainsAnyFlags(isSug ? Affix.Flags_ForbiddenWord_OnlyUpcase_NoSuggest : Affix.Flags_ForbiddenWord_OnlyUpcase))
                             {
                                 st.Destroy();
                                 return null;
@@ -808,86 +635,84 @@ public partial class WordList
                                 wordNum++;
                             }
 
-                            firstWordCompoundAcceptable =
+                            if (
                                 (
-                                    checkedPrefix
-                                    ||
-                                    (words?.CheckIfCurrentIsNotNull()).GetValueOrDefault()
-                                    ||
-                                    rv.ContainsFlag(Affix.CompoundFlag)
-                                    ||
+                                    !checkedPrefix
+                                    &&
+                                    !(words?.CheckIfCurrentIsNotNull()).GetValueOrDefault()
+                                    &&
+                                    !rv.ContainsAnyFlags(oldwordnum == 0 ? Affix.Flags_CompoundFlag_CompoundBegin : Affix.Flags_CompoundFlag_CompoundMiddle)
+                                    &&
                                     (
-                                        oldwordnum == 0
-                                            ? rv.ContainsFlag(Affix.CompoundBegin)
-                                            : (oldwordnum > 0 && rv.ContainsFlag(Affix.CompoundMiddle))
-                                    )
-                                    ||
-                                    (
-                                        huMovRule
-                                        && // LANG_hu section: spec. Hungarian rule
-                                        Affix.IsHungarian
-                                        && // XXX hardwired Hungarian dictionary codes
-                                        rv.Detail.ContainsAnyFlags(SpecialFlags.LetterF, SpecialFlags.LetterG, SpecialFlags.LetterH)
+                                        !huMovRule
+                                        || // LANG_hu section: spec. Hungarian rule
+                                        !Affix.IsHungarian
+                                        || // XXX hardwired Hungarian dictionary codes
+                                        !rv.ContainsAnyFlags(SpecialFlags.SetFGH)
                                     ) // END of LANG_hu section
                                 )
-                                && // test CHECKCOMPOUNDPATTERN conditions
-                                (
-                                    scpd == 0
-                                    ||
-                                    scpdPatternEntryCondition.IsZero
-                                    ||
-                                    rv.ContainsFlag(scpdPatternEntryCondition)
-                                )
-                                &&
+                                || // test CHECKCOMPOUNDPATTERN conditions
                                 (
                                     scpd != 0
-                                    ||
-                                    words is not null
-                                    ||
+                                    &&
+                                    scpdPatternEntryCondition.HasValue
+                                    &&
+                                    !rv.ContainsFlag(scpdPatternEntryCondition)
+                                )
+                                ||
+                                (
+                                    scpd == 0
+                                    &&
+                                    words is null
+                                    &&
                                     (
                                         (
-                                            !Affix.CheckCompoundTriple
-                                            || // test triple letters
-                                            i <= 0
-                                            ||
-                                            i >= word.Length
-                                            ||
-                                            word[i - 1] != word[i]
-                                            ||
+                                            Affix.CheckCompoundTriple
+                                            && // test triple letters
+                                            i > 0
+                                            &&
+                                            i < word.Length
+                                            &&
+                                            word[i - 1] == word[i]
+                                            &&
                                             (
-                                                (i < 2 || word[i - 1] != word[i - 2])
-                                                &&
-                                                (i + 1 >= word.Length || word[i - 1] != word[i + 1]) // may be word[i+1] == '\0'
+                                                (i >= 2 && word[i - 1] == word[i - 2])
+                                                ||
+                                                (i + 1 < word.Length && word[i - 1] == word[i + 1]) // may be word[i+1] == '\0'
                                             )
                                         )
-                                        &&
+                                        ||
                                         (
-                                            !Affix.CheckCompoundCase
-                                            ||
-                                            word.IsEmpty
-                                            ||
-                                            !CompoundCaseCheck(word, i)
+                                            Affix.CheckCompoundCase
+                                            &&
+                                            !word.IsEmpty
+                                            &&
+                                            CompoundCaseCheck(word, i)
                                         )
                                     )
-                                );
+                                )
+                            )
+                            {
+                                rv = null;
+                            }
                         }
                         else if (huMovRule && Affix.IsHungarian)
                         {
                             rv = AffixCheck(st.TerminatedSpan, default, CompoundOptions.Not);
 
-                            firstWordCompoundAcceptable =
+                            if(
                                 rv is not null
                                 // XXX hardwired Hungarian dic. codes
                                 &&
-                                (Suffix?.ContainsAnyContClass(SpecialFlags.LetterXLower, SpecialFlags.LetterPercent)).GetValueOrDefault();
-                        }
-                        else
-                        {
-                            firstWordCompoundAcceptable = false;
+                                (_suffix?.ContainsAnyContClass(SpecialFlags.SetXPercent)).GetValueOrDefault()
+                            )
+                            {
+                                rv = null;
+                            }
                         }
 
                         // first word is acceptable in compound words?
-                        if (firstWordCompoundAcceptable)
+                        if (rv is not null)
                         {
                             // first word is ok condition
                             if (Affix.IsHungarian)
@@ -897,24 +722,28 @@ public partial class WordList
 
                                 // - affix syllable num.
                                 // XXX only second suffix (inflections, not derivations)
-                                if (SuffixAppend is not null)
+                                if (_suffixAppend is not null)
                                 {
-                                    numSyllable -= GetSyllableReversed(SuffixAppend.AsSpan());
+                                    numSyllable -= GetSyllableReversed(_suffixAppend.AsSpan());
                                 }
-                                if (SuffixExtra)
+                                if (_suffixExtra)
                                 {
                                     numSyllable -= 1;
                                 }
 
                                 // + 1 word, if syllable number of the prefix > 1 (hungarian convention)
-                                if (Prefix is not null && GetSyllable(Prefix.Key.AsSpan()) > 1)
+                                if (_prefix is not null && GetSyllable(_prefix.Key.AsSpan()) > 1)
                                 {
                                     wordNum++;
                                 }
                             }
 
+#pragma warning disable IDE0007 // Use implicit type
+
                             // NEXT WORD(S)
-                            WordEntry rvFirst = rv!; // firstWordCompoundAcceptable ensures that rv is not null
+                            WordEntry rvFirst = rv;
+
+#pragma warning restore IDE0007 // Use implicit type
 
                             st[i] = ch;
 
@@ -942,7 +771,7 @@ public partial class WordList
                                 if (rv is not null)
                                 {
                                     // check FORCEUCASE
-                                    if (!EnumEx.HasFlag(info, SpellCheckResultType.OrigCap) && rv.ContainsFlag(Affix.ForceUpperCase))
+                                    if (!info.HasFlagEx(SpellCheckResultType.OrigCap) && rv.ContainsFlag(Affix.ForceUpperCase))
                                     {
                                         rv = null;
                                     }
@@ -976,7 +805,7 @@ public partial class WordList
                                     }
 
                                     // check forbiddenwords
-                                    if (ContainFlagsOrBlockSuggest(rv.Detail, isSug, Affix.ForbiddenWord, SpecialFlags.OnlyUpcaseFlag))
+                                    if (rv.ContainsAnyFlags(isSug ? Affix.Flags_ForbiddenWord_OnlyUpcase_NoSuggest : Affix.Flags_ForbiddenWord_OnlyUpcase))
                                     {
                                         st.Destroy();
                                         return null;
@@ -990,7 +819,7 @@ public partial class WordList
                                     if (
                                         rv is not null
                                         &&
-                                        rv.Detail.ContainsAnyFlags(Affix.CompoundFlag, Affix.CompoundEnd)
+                                        rv.ContainsAnyFlags(Affix.Flags_CompoundFlag_CompoundEnd)
                                         &&
                                         (
                                             !Affix.CompoundWordMax.HasValue
@@ -1047,28 +876,22 @@ public partial class WordList
                                 // perhaps second word has prefix or/and suffix
                                 ClearSuffixAndFlag();
 
+                                if (i < word.Length)
                                 {
-                                    var wordSubI = word.Slice(Math.Min(i, word.Length));
-                                    rv = (!onlycpdrule && Affix.CompoundFlag.HasValue && !wordSubI.IsEmpty)
-                                         ? AffixCheck(wordSubI, Affix.CompoundFlag, CompoundOptions.End)
+                                    rv = (!onlycpdrule && Affix.CompoundFlag.HasValue)
+                                         ? AffixCheck(word.Slice(i), Affix.CompoundFlag, CompoundOptions.End)
                                          : null;
 
                                     if (rv is null && Affix.CompoundEnd.HasValue && !onlycpdrule)
                                     {
                                         ClearSuffix();
                                         ClearPrefix();
-                                        if (!wordSubI.IsEmpty)
-                                        {
-                                            rv = AffixCheck(wordSubI, Affix.CompoundEnd, CompoundOptions.End);
-                                        }
+                                        rv = AffixCheck(word.Slice(i), Affix.CompoundEnd, CompoundOptions.End);
                                     }
 
                                     if (rv is null && Affix.CompoundRules.HasItems && words is not null)
                                     {
-                                        if (!wordSubI.IsEmpty)
-                                        {
-                                            rv = AffixCheck(wordSubI, default, CompoundOptions.End);
-                                        }
+                                        rv = AffixCheck(word.Slice(i), default, CompoundOptions.End);
 
                                         if (rv is not null && DefCompoundCheck(words.CreateIncremented(), rv.Detail, true))
                                         {
@@ -1079,14 +902,18 @@ public partial class WordList
                                         rv = null;
                                     }
                                 }
+                                else
+                                {
+                                    rv = null;
+                                }
 
-                                if(
+                                if (
                                     rv is not null
                                     &&
                                     (
                                         // check FORCEUCASE
                                         (
-                                            !EnumEx.HasFlag(info, SpellCheckResultType.OrigCap)
+                                            !info.HasFlagEx(SpellCheckResultType.OrigCap)
                                             &&
                                             rv.ContainsFlag(Affix.ForceUpperCase)
                                         )
@@ -1116,11 +943,7 @@ public partial class WordList
                                 }
 
                                 // check forbiddenwords
-                                if (
-                                    rv is not null
-                                    &&
-                                    ContainFlagsOrBlockSuggest(rv.Detail, isSug, Affix.ForbiddenWord, SpecialFlags.OnlyUpcaseFlag)
-                                )
+                                if ((rv?.ContainsAnyFlags(isSug ? Affix.Flags_ForbiddenWord_OnlyUpcase_NoSuggest : Affix.Flags_ForbiddenWord_OnlyUpcase)).GetValueOrDefault())
                                 {
                                     st.Destroy();
                                     return null;
@@ -1141,18 +964,18 @@ public partial class WordList
 
                                     // - affix syllable num.
                                     // XXX only second suffix (inflections, not derivations)
-                                    if (SuffixAppend is not null)
+                                    if (_suffixAppend is not null)
                                     {
-                                        numSyllable -= GetSyllableReversed(SuffixAppend.AsSpan());
+                                        numSyllable -= GetSyllableReversed(_suffixAppend.AsSpan());
                                     }
-                                    if (SuffixExtra)
+                                    if (_suffixExtra)
                                     {
                                         numSyllable -= 1;
                                     }
 
                                     // + 1 word, if syllable number of the prefix > 1 (hungarian
                                     // convention)
-                                    if (Prefix is not null && GetSyllable(Prefix.Key.AsSpan()) > 1)
+                                    if (_prefix is not null && GetSyllable(_prefix.Key.AsSpan()) > 1)
                                     {
                                         wordNum++;
                                     }
@@ -1162,17 +985,17 @@ public partial class WordList
 
                                     if (!string.IsNullOrEmpty(Affix.CompoundSyllableNum))
                                     {
-                                        if (SuffixFlag == SpecialFlags.LetterCLower)
+                                        if (_suffixFlag == SpecialFlags.LetterCLower)
                                         {
                                             numSyllable += 2;
                                         }
                                         else if (
-                                            SuffixFlag == SpecialFlags.LetterJ
+                                            _suffixFlag == SpecialFlags.LetterJ
                                             ||
                                             (
-                                                SuffixFlag == SpecialFlags.LetterI
-                                                && rv is not null
-                                                && rv.ContainsFlag(SpecialFlags.LetterJ)
+                                                _suffixFlag == SpecialFlags.LetterI
+                                                &&
+                                                (rv?.ContainsFlag(SpecialFlags.LetterJ)).GetValueOrDefault()
                                             )
                                         )
                                         {
@@ -1229,12 +1052,12 @@ public partial class WordList
                                 // perhaps second word is a compound word (recursive call)
                                 // (only if SPELL_COMPOUND_2 is not set and maxwordnum is not exceeded)
                                 if (
-                                    !info.HasFlag(SpellCheckResultType.Compound2)
+                                    !info.HasFlagEx(SpellCheckResultType.Compound2)
                                     &&
                                     (wordNum + 2) < maxwordnum
                                 )
                                 {
-                                    rv = CompoundCheck(st.TerminatedSpan.Slice(i), wordNum + 1, numSyllable, maxwordnum, words?.CreateIncremented(), rwords.CreateIncremented(), false, isSug, ref info, ref opLimiter);
+                                    rv = CompoundCheck(st.TerminatedSpan.Slice(i), wordNum + 1, numSyllable, maxwordnum, words?.CreateIncremented(), rwords.CreateIncremented(), false, isSug, info, ref opLimiter);
 
                                     if (
                                         rv is not null
@@ -1355,7 +1178,7 @@ public partial class WordList
                         st[i] = ch;
                     }
                 }
-                while (Affix.CompoundRules.HasItems && oldwordnum == 0 && IntEx.InversePostfixIncrement(ref onlycpdrule)); // end of onlycpd loop
+                while (Affix.CompoundRules.HasItems && oldwordnum == 0 && inversePostfixIncrement(ref onlycpdrule)); // end of onlycpd loop
             }
 
             st.Destroy();
@@ -1363,6 +1186,167 @@ public partial class WordList
 
             static bool equalsOrdinalLimited(ReadOnlySpan<char> a, ReadOnlySpan<char> b, int lengthLimit) =>
                 a.Limit(lengthLimit).EqualsOrdinal(b.Limit(lengthLimit));
+
+            static bool inversePostfixIncrement(ref bool b)
+            {
+                // This is a strange implementation of a postfix increment to remove the use of an int in these cases.
+
+                if (b)
+                {
+                    return false;
+                }
+
+                b = true;
+                return true;
+            }
+        }
+
+        private WordEntry? CompoundCheckWordSearch_AffixCheckCompoundFlag(ReadOnlySpan<char> st, bool huMovRule)
+        {
+            var movCompoundOptions = huMovRule ? CompoundOptions.Other : CompoundOptions.Begin;
+
+            var rv = PrefixCheck(st, movCompoundOptions, Affix.CompoundFlag);
+
+            if (rv is null)
+            {
+                rv = SuffixCheck(st, AffixEntryOptions.None, null, FlagValue.Zero, Affix.CompoundFlag, movCompoundOptions);
+
+                if (rv is null && Affix.CompoundMoreSuffixes)
+                {
+                    rv = SuffixCheckTwoSfx(st, AffixEntryOptions.None, null, Affix.CompoundFlag);
+                }
+
+                if (
+                    rv is not null
+                    &&
+                    !huMovRule
+                    &&
+                    (_suffix?.ContainsAnyContClass(Affix.Flags_CompoundForbid_CompoundEnd)).GetValueOrDefault()
+                )
+                {
+                    rv = null;
+                }
+            }
+
+            return rv;
+        }
+
+        private WordEntry? CompoundCheckWordSearch_AffixCheckCompoundBegin(ReadOnlySpan<char> st, bool huMovRule)
+        {
+            var movCompoundOptions = huMovRule ? CompoundOptions.Other : CompoundOptions.Begin;
+
+            var rv = SuffixCheck(st, AffixEntryOptions.None, null, default, Affix.CompoundBegin, movCompoundOptions);
+
+            if (rv is null)
+            {
+                if (Affix.CompoundMoreSuffixes)
+                {
+                    rv = SuffixCheckTwoSfx(st, AffixEntryOptions.None, null, Affix.CompoundBegin);
+                }
+
+                rv ??= PrefixCheck(st, movCompoundOptions, Affix.CompoundBegin);
+            }
+
+            return rv;
+        }
+
+        private WordEntry? CompoundCheckWordSearch_AffixCheckCompoundMiddle(ReadOnlySpan<char> st, bool huMovRule)
+        {
+            var movCompoundOptions = huMovRule ? CompoundOptions.Other : CompoundOptions.Begin;
+
+            var rv = SuffixCheck(st, AffixEntryOptions.None, null, default, Affix.CompoundMiddle, movCompoundOptions);
+
+            if (rv is null)
+            {
+                if (Affix.CompoundMoreSuffixes)
+                {
+                    rv = SuffixCheckTwoSfx(st, AffixEntryOptions.None, null, Affix.CompoundMiddle);
+                }
+
+                rv ??= PrefixCheck(st, movCompoundOptions, Affix.CompoundMiddle);
+            }
+
+            return rv;
+        }
+
+        private readonly WordEntryDetail? CompoundCheckWordSearch_OnlyCpdRule(
+            WordEntryDetail[] searchEntryDetails,
+            FlagValue scpdPatternEntryCondition,
+            ref IncrementalWordList? words,
+            IncrementalWordList rwords)
+        {
+            foreach (var searchEntryDetail in searchEntryDetails)
+            {
+                if (!searchEntryDetail.ContainsFlag(Affix.NeedAffix))
+                {
+                    // NOTE: do not reorder DefCompoundCheck calls or other conditions due to side effects
+
+                    bool defCpdCheck;
+                    if (words is not null)
+                    {
+                        defCpdCheck = DefCompoundCheck(words, searchEntryDetail, false);
+                    }
+                    else if (DefCompoundCheck(rwords, searchEntryDetail, false))
+                    {
+                        words = rwords;
+                        defCpdCheck = true;
+                    }
+                    else
+                    {
+                        defCpdCheck = false;
+                    }
+
+                    // NOTE: do not reorder DefCompoundCheck calls or other conditions due to side effects
+
+                    if (defCpdCheck && (scpdPatternEntryCondition.IsZero || searchEntryDetail.ContainsFlag(scpdPatternEntryCondition)))
+                    {
+                        return searchEntryDetail;
+                    }
+                }
+            }
+
+            return null;
+        }
+
+        private readonly WordEntryDetail? CompoundCheckWordSearch_NormalNoWordList(
+            WordEntryDetail[] searchEntryDetails,
+            FlagValue scpdPatternEntryCondition,
+            FlagValue compoundPart)
+        {
+            foreach (var searchEntryDetail in searchEntryDetails)
+            {
+                if (
+                    !searchEntryDetail.ContainsFlag(Affix.NeedAffix)
+                    &&
+                    searchEntryDetail.Flags.ContainsAny(Affix.CompoundFlag, scpdPatternEntryCondition, compoundPart)
+                )
+                {
+                    return searchEntryDetail;
+                }
+            }
+
+            return null;
+        }
+
+        private readonly WordEntryDetail? CompoundCheckWordSearch_NormalWithExistingWordList(
+            WordEntryDetail[] searchEntryDetails,
+            FlagValue scpdPatternEntryCondition)
+        {
+            foreach (var searchEntryDetail in searchEntryDetails)
+            {
+                if (
+                    !searchEntryDetail.ContainsFlag(Affix.NeedAffix)
+                    &&
+                    searchEntryDetail.ContainsFlag(Affix.CompoundBegin)
+                    &&
+                    (scpdPatternEntryCondition.IsZero || searchEntryDetail.ContainsFlag(scpdPatternEntryCondition))
+                )
+                {
+                    return searchEntryDetail;
+                }
+            }
+
+            return null;
         }
 
         /// <summary>
@@ -1382,15 +1366,12 @@ public partial class WordList
                     ClearSuffix();
                     ClearPrefix();
 
-                    if (rv is null)
-                    {
-                        rv =
-                            // if still not found check all two-level suffixes
-                            SuffixCheckTwoSfx(word, AffixEntryOptions.None, null, needFlag)
-                            ??
-                            // if still not found check all two-level prefixes
-                            PrefixCheckTwoSfx(word, CompoundOptions.Not, needFlag);
-                    }
+                    rv ??=
+                        // if still not found check all two-level suffixes
+                        SuffixCheckTwoSfx(word, AffixEntryOptions.None, null, needFlag)
+                        ??
+                        // if still not found check all two-level prefixes
+                        PrefixCheckTwoSfx(word, CompoundOptions.Not, needFlag);
                 }
             }
 
@@ -1523,7 +1504,7 @@ public partial class WordList
                     // if CrossProduct is allowed, try again but now
                     // cross checked combined with a suffix
 
-                    if (EnumEx.HasFlag(pe.Options, AffixEntryOptions.CrossProduct) && (inCompound != CompoundOptions.Begin))
+                    if (pe.Options.HasFlagEx(AffixEntryOptions.CrossProduct) && (inCompound != CompoundOptions.Begin))
                     {
                         // find hash entry of root word
                         if (SuffixCheckTwoSfx(tmpword, AffixEntryOptions.CrossProduct, pe, needFlag) is { } he)
@@ -1718,9 +1699,9 @@ public partial class WordList
                 if (Affix.ContClasses.Contains(affix.AFlag))
                 {
                     rv = CheckTwoSfx(affix, word, sfxopts, pfx, needflag);
-                    if (rv is not null && Suffix is not null)
+                    if (rv is not null && _suffix is not null)
                     {
-                        SetSuffixFlag(Suffix.AFlag);
+                        SetSuffixFlag(_suffix.AFlag);
                         if (!affix.ContClass.HasItems)
                         {
                             SetSuffixAppend(affix.Key);
@@ -1734,21 +1715,23 @@ public partial class WordList
             return null;
         }
 
-        private WordEntry? LookupFirst(ReadOnlySpan<char> word) => WordList.FindFirstEntryByRootWord(word);
+        private readonly WordEntry? LookupFirst(ReadOnlySpan<char> word) => WordList.FindFirstEntryByRootWord(word);
 
-        private WordEntry? LookupFirst(string word) => WordList.FindFirstEntryByRootWord(word);
+        private readonly WordEntry? LookupFirst(string word) => WordList.FindFirstEntryByRootWord(word);
 
-        public bool TryLookupDetails(ReadOnlySpan<char> word, out string actualKey, out WordEntryDetail[] details) => WordList.EntriesByRoot.TryGetValue(word, out actualKey, out details);
+        public readonly bool TryLookupDetails(ReadOnlySpan<char> word, out string actualKey, out WordEntryDetail[] details) =>
+            WordList.EntriesByRoot.TryGetValue(word, out actualKey, out details);
 
-        public bool TryLookupDetails(string word, out WordEntryDetail[] details) => WordList.EntriesByRoot.TryGetValue(word, out details);
+        public readonly bool TryLookupDetails(string word, out WordEntryDetail[] details) =>
+            WordList.EntriesByRoot.TryGetValue(word, out details);
 
         /// <summary>
         /// Compound check patterns.
         /// </summary>
-        private bool DefCompoundCheck(IncrementalWordList words, in WordEntryDetail rv, bool all)
+        private readonly bool DefCompoundCheck(IncrementalWordList words, in WordEntryDetail rv, bool all)
         {
             // has the last word COMPOUNDRULE flag?
-            if (Affix.CompoundRules.EntryContainsRuleFlags(rv))
+            if (Affix.CompoundRules.EntryContainsRuleFlags(rv.Flags))
             {
                 words.SetCurrent(rv);
                 if (Affix.CompoundRules.CompoundCheck(words, all))
@@ -1807,7 +1790,7 @@ public partial class WordList
         /// <summary>
         /// Calculate number of syllable for compound-checking.
         /// </summary>
-        private int GetSyllable(ReadOnlySpan<char> word)
+        private readonly int GetSyllable(ReadOnlySpan<char> word)
         {
             var num = 0;
             if (Affix.CompoundMaxSyllable != 0 && Affix.CompoundVowels.HasItems)
@@ -1827,7 +1810,7 @@ public partial class WordList
         /// Calculate number of syllable for compound-checking.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private int GetSyllableReversed(ReadOnlySpan<char> word)
+        private readonly int GetSyllableReversed(ReadOnlySpan<char> word)
         {
             // Because the code is effectively the same, forward and reverse searches can use the same algorithm
             return GetSyllable(word);
@@ -1948,7 +1931,7 @@ public partial class WordList
                     // if aeXPRODUCT is allowed, try again but now
                     // ross checked combined with a suffix
 
-                    if (EnumEx.HasFlag(affix.Options, AffixEntryOptions.CrossProduct))
+                    if (affix.Options.HasFlagEx(AffixEntryOptions.CrossProduct))
                     {
                         if (SuffixCheck(tmpword, AffixEntryOptions.CrossProduct, affix, default, needFlag, inCompound) is { } he)
                         {
@@ -1962,12 +1945,12 @@ public partial class WordList
             return null;
         }
 
-        private WordEntry? CheckWordSuffix(SuffixEntry affix, ReadOnlySpan<char> word, AffixEntryOptions optFlags, PrefixEntry? pfx, FlagValue cclass, FlagValue needFlag, FlagValue badFlag)
+        private readonly WordEntry? CheckWordSuffix(SuffixEntry affix, ReadOnlySpan<char> word, AffixEntryOptions optFlags, PrefixEntry? pfx, FlagValue cclass, FlagValue needFlag, FlagValue badFlag)
         {
             // if this suffix is being cross checked with a prefix
             // but it does not support cross products skip it
 
-            var optFlagsHasCrossProduct = EnumEx.HasFlag(optFlags, AffixEntryOptions.CrossProduct);
+            var optFlagsHasCrossProduct = optFlags.HasFlagEx(AffixEntryOptions.CrossProduct);
             if (
                 (
                     optFlagsHasCrossProduct
@@ -1975,7 +1958,7 @@ public partial class WordList
                     (
                         pfx is null // enabled by prefix is impossible
                         ||
-                        !EnumEx.HasFlag(affix.Options, AffixEntryOptions.CrossProduct)
+                        !affix.Options.HasFlagEx(AffixEntryOptions.CrossProduct)
                     )
                 )
                 ||
@@ -2069,7 +2052,7 @@ public partial class WordList
             // if this suffix is being cross checked with a prefix
             // but it does not support cross products skip it
 
-            if (EnumEx.HasFlag(optflags, AffixEntryOptions.CrossProduct) && !EnumEx.HasFlag((AffixEntryOptions)se.Options, AffixEntryOptions.CrossProduct))
+            if (optflags.HasFlagEx(AffixEntryOptions.CrossProduct) && !se.Options.HasFlagEx(AffixEntryOptions.CrossProduct))
             {
                 return null;
             }
@@ -2120,10 +2103,10 @@ public partial class WordList
         }
 
         private bool CandidateCheck(string word) =>
-            WordList.ContainsEntriesForRootWord(word) || AffixCheck(word.AsSpan(), default(FlagValue), CompoundOptions.Not) is not null;
+            WordList.ContainsEntriesForRootWord(word) || AffixCheck(word.AsSpan(), default, CompoundOptions.Not) is not null;
 
         private bool CandidateCheck(ReadOnlySpan<char> word) =>
-            WordList.ContainsEntriesForRootWord(word) || AffixCheck(word, default(FlagValue), CompoundOptions.Not) is not null;
+            WordList.ContainsEntriesForRootWord(word) || AffixCheck(word, default, CompoundOptions.Not) is not null;
 
         /// <summary>
         /// Make a copy of <paramref name="src"/> and returns it
@@ -2140,7 +2123,7 @@ public partial class WordList
         /// set the capitalization type (<paramref name="capType"/>) and
         /// return the length of the "cleaned" (and UTF-8 encoded) word
         /// </remarks>
-        public string CleanWord2(string src, out CapitalizationType capType, out int abbv)
+        public readonly string CleanWord2(string src, out CapitalizationType capType, out int abbv)
         {
             if (Affix.IgnoredChars.HasItems)
             {
@@ -2170,7 +2153,7 @@ public partial class WordList
             return src;
         }
 
-        public string CleanWord2(ReadOnlySpan<char> src, out CapitalizationType capType, out int abbv)
+        public readonly string CleanWord2(ReadOnlySpan<char> src, out CapitalizationType capType, out int abbv)
         {
             if (Affix.IgnoredChars.HasItems)
             {
